FIREFLY-686: moved unit conversion logic out into its own file
